Add dosing to eventTable

This adds a dosing event to the event table. This is provided for piping syntax through magrittr

add.dosing(eventTable, dose, nbr.doses = 1L, dosing.interval = 24,
  dosing.to = 1L, rate = NULL, amount.units = NA_character_,
  start.time = 0, do.sampling = FALSE, time.units = NA_character_, ...)

Arguments

eventTable
eventTable object

dose
numeric scalar, dose amount in amount.units;

nbr.doses
integer, number of doses;

dosing.interval
required numeric scalar, time between doses in time.units, defaults to 24 of time.units="hours";

dosing.to
integer, compartment the dose goes into (first compartment by default);

rate
for infusions, the rate of infusion (default is NULL, for bolus dosing;

amount.units
optional string indicating the dosing units. Defaults to NA to indicate as per the original EventTable definition.

start.time
required dosing start time;

do.sampling
logical, should observation sampling records be added at the dosing times? Defaults to FALSE.

time.units
optional string indicating the time units. Defaults to "hours" to indicate as per the original EventTable definition.

...
Other parameters (ignored)

Value

eventTable with updated dosing (note the event table will be updated anyway)

See also