Return the md5 of an RxODE object or file

This md5 is based on the model and possibly the extra c code supplied for the model. In addition the md5 is based on syntax options, compiled RxODE library md5, and the RxODE version/repository.

rxMd5(model, extraC = NULL, ...)

Arguments

model
This is the ODE model specification. It can be:

a string containing the set of ordinary differential equations (ODE) and other expressions defining the changes in the dynamic system.

a file name where the ODE system equation is contained

An ODE expression enclosed in {}

(see also the filename argument). For details, see the sections “Details” and “RxODE Syntax” below.

extraC
Extra c code to include in the model. This can be useful to specify functions in the model. These C functions should usually take double precision arguments, and return double precision values.

...
ignored arguments

Value

If this is a RxODE object, return a named list:

file_md5

is the model's file's md5

parsed_md5

is the parsed model's file's md5.

Otherwise return the md5 based on the arguments provided