Use Shiny to help develop an RxODE model

Use Shiny to help develop an RxODE model

rxShiny(object, params = c(), events = NULL, inits = c(), ...,
  data = data.frame())

# S3 method for rxSolve
rxShiny(object, params = NULL, events = NULL,
  inits = c(), ..., data = data.frame())

# S3 method for default
rxShiny(object = NULL, params = c(), events = NULL,
  inits = c(), ..., data = data.frame())

Arguments

object
A RxODE family of objects. If not supplied a 2-compartment indirect effect model is used. If it is supplied, use the model associated with the RxODE object for the model exploration.

params
Initial parameters for model

events
Event information (currently ignored)

inits
Initial estimates for model

...
Other arguments passed to rxShiny. Currently doesn't do anything.

data
Any data that you would like to plot. If the data has a time variable as well as a compartment or calculated variable that matches the RxODE model, the data will be added to the plot of a specific compartment or calculated variable.

Value

Nothing; Starts a shiny server