Simulate Parameters from a Theta/Omega specification

Simulate Parameters from a Theta/Omega specification

rxSimThetaOmega(params = NULL, omega = NULL, omegaDf = NULL,
  omegaIsChol = FALSE, nSub = 1L, thetaMat = NULL, thetaDf = NULL,
  thetaIsChol = FALSE, nStud = 1L, sigma = NULL, sigmaDf = NULL,
  sigmaIsChol = FALSE, nCoresRV = 1L, nObs = 1L, dfSub = 0, dfObs = 0,
  simSubjects = TRUE)

Arguments

params
Named Vector of RxODE model parameters

omega
Named omega matrix.

omegaDf
The degrees of freedom of a t-distribution for simulation. By default this is NULL which is equivalent to Inf degrees, or to simulate from a normal distribution instead of a t-distribution.

omegaIsChol
Indicates if the omega supplied is a Cholesky decomposed matrix instead of the traditional symmetric matrix.

nSub
Number between subject variabilities (ETAs) simulated for every realization of the parameters.

thetaMat
Named theta matrix.

thetaDf
The degrees of freedom of a t-distribution for simulation. By default this is NULL which is equivalent to Inf degrees, or to simulate from a normal distribution instead of a t-distribution.

thetaIsChol
Indicates if the theta supplied is a Cholesky decomposed matrix instead of the traditional symmetric matrix.

nStud
Number virtual studies to characterize uncertainty in estimated parameters.

sigma
Matrix for residual variation. Adds a "NA" value for each of the indivdual parameters, residuals are updated after solve is completed.

sigmaDf
Degrees of freedom of the sigma t-distribution. By default it is equivalent to Inf, or a normal distribution.

sigmaIsChol
Boolean indicating if the sigma is in the Cholesky decomposition instead of a symmetric covariance

nCoresRV
Number of cores used for the simulation of the sigma variables. By default this is 1. This uses the package rmvn and rmvt. To reproduce the results you need to run on the same platform with the same number of cores. This is the reason this is set to be one, regardless of what the number of cores are used in threaded ODE solving.

nObs
Number of observations to simulate (with sigma matrix)

dfSub
Degrees of freedom to sample the between subject variaiblity matrix from the inverse Wishart distribution (scaled) or scaled inverse chi squared distribution.

dfObs
Degrees of freedom to sample the unexplained variaiblity matrix from the inverse Wishart distribution (scaled) or scaled inverse chi squared distribution.

simSubjects
boolean indicated RxODE should simulate subjects in studies (TRUE, default) or studies (FALSE)