Simulate Weibull variable from threefry generator

Care should be taken with this method not to encounter the birthday problem, described https://www.johndcook.com/blog/2016/01/29/random-number-generator-seed-mistakes/. Since the sitmo threefry, this currently generates one random deviate from the uniform distribution to seed the engine threefry and then run the code.

rxweibull(shape, scale = 1, n = 1L, ncores = 1L)

Arguments

shape
shape and scale parameters, the latter defaulting to 1.

scale
shape and scale parameters, the latter defaulting to 1.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

ncores
Number of cores for the simulation

rxnorm simulates using the threefry sitmo generator; rxnormV uses the vandercorput generator

Value

Weibull random deviates

Details

Therefore, a simple call to the random number generated followed by a second call to random number generated may have identical seeds. As the number of random number generator calls are increased the probability that the birthday problem will increase.

The key to avoid this problem is to run all simulations in the RxODE environment once (therefore one seed or series of seeds for the whole simulation) or to pre-generate all random variables used for the simulation.

Also care should be made that the computer you will be running on can run the same number of cores as you are running so they can reproduce your results.

Examples

# \donttest{

## Use threefry engine

rxweibull(shape=1, scale=4, n=10) # with rxweibull you have to explicitly state n
#>  [1] 0.3486622 2.9635242 9.2460350 4.3840779 5.9624172 1.7138640 0.7363841
#>  [8] 1.2759314 1.0791512 2.9407383
rxweibull(shape=1, scale=4, n=10, ncores=2) # You can parallelize the simulation using openMP
#>  [1]  2.5187157  1.0679777  1.1732240  3.8040546  1.5357556  0.1385729
#>  [7]  0.3315010 10.8155110  8.4266018  3.1755766

rxweibull(3)
#> [1] 1.203126

## This example uses `rxweibull` directly in the model

rx <- RxODE({
  a = rxweibull(1,3)
})
#>  

et <- et(1,id=1:2)

s <- rxSolve(rx,et)
#> Warning: thread safe method, but results may depend on system/load, using 1 core (can change with `cores=`)

# }